audit_log/requirements.txt
a2wsgi==1.10.2
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
connexion==3.0.6
Flask==3.0.2
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kazoo==2.5.0
MarkupSafe==2.1.5
pykafka==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
swagger_ui_bundle==1.1.0
tabulate==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
